Avoid compile warning that only happens on macOS

+++ src/Private/Renderer/Utility/DebugPanels.cpp
@@ -134,7 +134,7 @@ namespace Renderer::Debug
             [](const AllocatedBuffer& buf, int last_column)
             {
                 ImGui::TableSetColumnIndex(last_column + 1);
-                ImGui::Text("%zu bytes", buf.allocation_info.size);
+                ImGui::Text("%zu bytes", (size_t)buf.allocation_info.size);
                 ImGui::TableSetColumnIndex(last_column + 2);
                 ImGui::Text("0x%p", buf.buffer);
             },
